requirements.txt
absl-py==1.0.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
autograd==1.3
autograd-gamma==0.5.0
backcall==0.2.0
batchgenerators==0.24
beautifulsoup4==4.11.1
bleach==4.1.0
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.4
cvxopt==1.3.0
cvxpy==1.1.20
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
easydict==1.9
ecos==2.0.10
einops==0.4.1
entrypoints==0.3
et-xmlfile==1.1.0
feather-format==0.4.1
formulaic==0.2.4
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
gast==0.5.3
glmnet-py==0.1.0b2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googledrivedownloader==0.4
graphviz==0.8.4
grpcio==1.48.1
h5py==3.1.0
humanize==3.14.0
idna==3.3
idna-ssl==1.1.0
imageio==2.13.3
importlib-metadata==3.10.1
interface-meta==1.2.4
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282412546/work
lifelines==0.26.3
linecache2==1.0.0
llvmlite==0.36.0
loss-landscapes==3.0.6
lxml==4.9.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
MedPy==0.4.0
menpo==0.11.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.2.0
multivolumefile==0.2.3
mxnet==1.9.1
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nibabel==3.2.1
nmslib==2.1.1
notebook==6.4.5
numba==0.53.1
numexpr==2.8.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
nystrom-attention==0.0.11
oauthlib==3.2.0
olefile==0.46
onnx==1.11.0
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
openslide-python==1.1.2
osqp==0.6.2.post5
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prettytable==2.5.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.1
psutil==5.9.1
ptyprocess==0.7.0
py7zr==0.19.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pybind11==2.6.1
pycox==0.2.3
pycparser==2.20
pycryptodomex==3.15.0
pyDeprecate==0.3.1
pydicom==2.3.1
Pygments==2.10.0
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyppmd==0.18.3
pyrsistent==0.18.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-lightning==1.5.10
pytorch-metric-learning==1.6.0
pytz==2021.3
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
pyzstd==0.15.3
qdldl==0.1.5.post2
qtconsole==5.1.1
QtPy==1.11.2
rdflib==5.0.0
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-survival==0.14.0
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
scs==3.2.0
seaborn==0.11.2
Send2Trash==1.8.0
SimpleITK==1.2.4
six @ file:///tmp/build/80754af9/six_1623709665295/work
soupsieve==2.3.2.post1
statsmodels==0.12.2
Surface-Distance-Based-Measures @ file:///data_local/HJW/Seg/surface-distance
tables==3.7.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.6.7
torch==1.10.0
torch-geometric==2.0.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torchaudio==0.10.0
torchfile==0.1.0
torchio==0.18.76
torchmetrics==0.8.2
torchtuples==0.2.2
torchvision==0.11.1
tornado @ file:///tmp/build/80754af9/tornado_1606942266872/work
tqdm==4.62.3
traceback2==1.4.0
traitlets==4.3.3
typing-extensions==3.10.0.2
unittest2==1.1.0
urllib3==1.26.7
visdom==0.1.8.9
volumentations==0.1.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
widgetsnbextension==3.5.1
wrapt==1.13.3
yacs==0.1.8
yarl==1.7.2
zipp==3.6.0
